Initial contribution of modelserver-node
Adds the @eclipse-emfcloud/modelserver-node package.
This is a framework for building a node.js server that acts
as a façade for the EMF.Cloud Model Server (Java) server.
The modelserver-node server provides a Typescript (node.js)
environment for "business logic" customizing the Model Server
API via an extensible architecture in which plug-ins can
contribute

- providers of custom command implementations, optionally
  using a new transaction protocol that is private between the
  node.js layer and the Java server
- providers of custom validation rules
- custom API endpoints

Several Model Server API endpoints (in the v2 API) are
intercepted and delegated to provider plug-ins. Other
endpoints (including v1 API) are simply relayed to the
upstream Java server.

Contributed on behalf of STMicroelectronics.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>