Fix #34: Provide initial validation state in subscriptions
Add a listener API for clients such as the ValidationManager to
learn when subscriptions are added/removed. Update the
ValidationManager to use this event to report the initial
validation state to a new subscriber.

Some additional fixes are included in the SubscriptionManager
to support this and similar use cases in subclasses:
- export the SubscriptionQuery type used in public/protected
  method signatures
- provide convenient access to subscribers, including filtering
  on subscription options
- provide API for sending messages to subscribers

Contributed on behalf of STMicroelectronics.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>