Model Service to abstract model server delegation in plug-ins
Add a ModelService to locally proxy the Model Server API
for plug-in extensions to use.

Refactor the current implementation of edit endpoint extensions
and the transaction API for reuse in this new service.

Update dependencies to resolve TypeScript language version
mismatch in ESLint.

Work around issue in subscription URL processing in the
model server client delegate uncovered in development of
Mocha tests.

Resolves #46

Contributed on behalf of STMicroelectronics.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>