Crash when subscribing to non-existent model with live validation (#52)
The server intercepts subscriptions that have the
live validation option set and, when it is, initializes
the validation state by performing a validation.
The first step in this is to retrieve the model to
validate. If that fails with an exception, then the
subscription event handler bombs with the
unhandled promise rejection and (under Node 16
or later) the server process dies.

Improve the robustness of this live validation
initialization to trap and log the error. The
upstream server will take care of telling the
client application that the subscription failed.

Fixes #51.

Contributed on behalf of STMicroelectronics.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>